% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\docType{package}
\name{gepafer-package}
\alias{gepafer}
\alias{gepafer-package}
\title{gepafer: Google Encoded Polyline Algorithm Format}
\description{
Encode and decode the Google Encoded Polyline Algorithm Format.
}
\author{
\strong{Maintainer}: Matthieu Viry \email{matthieu.viry@cnrs.fr} (\href{https://orcid.org/0000-0002-0693-8556}{ORCID})

}
